+++ src/LASRapi/api.h
@@ -101,6 +101,8 @@ public:
   void set_progress(bool b) { opt_progress = b; };
   void set_chunk(double val) { if(val> 0) opt_chunk = val; };
   void set_profile_file(const std::string& path) { opt_profiling_file = path; };
+  void set_progress_file(const std::string& path) { opt_progress_file = path; };
+  void set_log_file(const std::string& path) { opt_log_file = path; };
   void set_noprocess(const std::vector<bool>&);
 
   bool has_reader() const;
@@ -125,9 +127,11 @@ private:
   bool opt_verbose = false;
   std::vector<bool> opt_noprocess;
   std::string opt_profiling_file = "";
+  std::string opt_progress_file = "";
+  std::string opt_log_file = "";
 };
 
-ReturnType execute(const std::string& config_file, const std::string& async_communication_file = "");
+ReturnType execute(const std::string& config_file);
 PipelineInfo pipeline_info(const std::string& config_file);
 
 Pipeline add_attribute(std::string data_type, std::string name, std::string description, double scale = 1, double offset = 0);